+++ common/spaces/http_space_primitives.ts
@@ -22,11 +22,8 @@ export class HttpSpacePrimitives implements SpacePrimitives {
     }
 
     const result = await fetch(url, { ...options });
-    if (result.headers.get("X-Status")) {
-      (result as any).status = +result.headers.get("X-Status")!;
-    }
     if (
-      result.status === 401
+      this.getRealStatus(result) === 401
     ) {
       // Invalid credentials, reloading the browser should trigger authentication
       console.log("Going to redirect after", url);
@@ -36,13 +33,20 @@ export class HttpSpacePrimitives implements SpacePrimitives {
     return result;
   }
 
+  getRealStatus(r: Response) {
+    if (r.headers.get("X-Status")) {
+      return +r.headers.get("X-Status")!;
+    }
+    return r.status;
+  }
+
   async fetchFileList(): Promise<FileMeta[]> {
     const resp = await this.authenticatedFetch(this.url, {
       method: "GET",
     });
 
     if (
-      resp.status === 200 &&
+      this.getRealStatus(resp) === 200 &&
       this.expectedSpacePath &&
       resp.headers.get("X-Space-Path") !== this.expectedSpacePath
     ) {
@@ -63,7 +67,7 @@ export class HttpSpacePrimitives implements SpacePrimitives {
         method: "GET",
       },
     );
-    if (res.status === 404) {
+    if (this.getRealStatus(res) === 404) {
       throw new Error(`Not found`);
     }
     return {
@@ -104,7 +108,7 @@ export class HttpSpacePrimitives implements SpacePrimitives {
         method: "DELETE",
       },
     );
-    if (req.status !== 200) {
+    if (this.getRealStatus(req) !== 200) {
       throw Error(`Failed to delete file: ${req.statusText}`);
     }
   }
@@ -116,7 +120,7 @@ export class HttpSpacePrimitives implements SpacePrimitives {
         method: "OPTIONS",
       },
     );
-    if (res.status === 404) {
+    if (this.getRealStatus(res) === 404) {
       throw new Error(`Not found`);
     }
     return this.responseToMeta(name, res);