+++ plugs/directive/query_directive.ts
@@ -5,6 +5,7 @@ import { renderTemplate } from "./util.ts";
 import { parseQuery } from "./parser.ts";
 import { jsonToMDTable } from "./util.ts";
 import { ParseTree } from "$sb/lib/tree.ts";
+import { folderName, resolve } from "$sb/lib/path.ts";
 
 export async function queryDirectiveRenderer(
   _directive: string,
@@ -33,8 +34,13 @@ export async function queryDirectiveRenderer(
   } else if (results.length === 1) {
     // console.log("Parsed query", parsedQuery);
     if (parsedQuery.render) {
-      const rendered = await renderTemplate(
+      const absoluteRenderPath = resolve(
+        folderName(pageName),
         parsedQuery.render,
+      );
+      const rendered = await renderTemplate(
+        pageName,
+        absoluteRenderPath,
         results[0],
       );
       return rendered.trim();