+++ web/cm_plugins/wiki_link.ts
@@ -1,5 +1,6 @@
 import { pageLinkRegex } from "../../common/markdown_parser/parser.ts";
 import { ClickEvent } from "../../plug-api/app_event.ts";
+import { folderName, resolve } from "../../plug-api/lib/path.ts";
 import { Decoration, syntaxTree } from "../deps.ts";
 import { Editor } from "../editor.tsx";
 import {
@@ -33,8 +34,13 @@ export function cleanWikiLinkPlugin(editor: Editor) {
         if (page.includes("@")) {
           cleanPage = page.split("@")[0];
         }
+        const resolvedPage = resolve(
+          folderName(editor.currentPage!),
+          cleanPage,
+        );
+        // console.log("Resolved page", resolvedPage);
         for (const pageMeta of allPages) {
-          if (pageMeta.name === cleanPage) {
+          if (pageMeta.name === resolvedPage) {
             pageExists = true;
             break;
           }
@@ -76,8 +82,10 @@ export function cleanWikiLinkPlugin(editor: Editor) {
             widget: new LinkWidget(
               {
                 text: linkText,
-                title: pageExists ? `Navigate to ${page}` : `Create ${page}`,
-                href: `/${page}`,
+                title: pageExists
+                  ? `Navigate to ${resolvedPage}`
+                  : `Create ${resolvedPage}`,
+                href: `/${resolvedPage}`,
                 cssClass: pageExists
                   ? "sb-wiki-link-page"
                   : "sb-wiki-link-page-missing",