+++ web/editor.tsx
@@ -138,6 +138,7 @@ import { FallbackSpacePrimitives } from "../common/spaces/fallback_space_primiti
 import { syncSyscalls } from "./syscalls/sync.ts";
 import { FilteredSpacePrimitives } from "../common/spaces/filtered_space_primitives.ts";
 import { CollabManager } from "./collab_manager.ts";
+import { folderName, relativePath } from "../plug-api/lib/path.ts";
 
 const frontMatterRegex = /^---\n(([^\n]|\n)*?)---\n/;
 
@@ -242,15 +243,8 @@ export class Editor {
       true,
     );
 
-    const plugSpacePrimitives = new PlugSpacePrimitives(
-      // Using fallback space primitives here to allow (by default) local reads to "fall through" to HTTP when files aren't synced yet
-      new FallbackSpacePrimitives(
-        new IndexedDBSpacePrimitives(
-          `${dbPrefix}_space`,
-          globalThis.indexedDB,
-        ),
-        this.remoteSpacePrimitives,
-      ),
+    const plugSpaceRemotePrimitives = new PlugSpacePrimitives(
+      this.remoteSpacePrimitives,
       namespaceHook,
     );
 
@@ -258,7 +252,14 @@ export class Editor {
     const localSpacePrimitives = new FilteredSpacePrimitives(
       new FileMetaSpacePrimitives(
         new EventedSpacePrimitives(
-          plugSpacePrimitives,
+          // Using fallback space primitives here to allow (by default) local reads to "fall through" to HTTP when files aren't synced yet
+          new FallbackSpacePrimitives(
+            new IndexedDBSpacePrimitives(
+              `${dbPrefix}_space`,
+              globalThis.indexedDB,
+            ),
+            plugSpaceRemotePrimitives,
+          ),
           this.eventHook,
         ),
         indexSyscalls,
@@ -279,12 +280,16 @@ export class Editor {
 
     this.syncService = new SyncService(
       localSpacePrimitives,
-      this.remoteSpacePrimitives,
+      plugSpaceRemotePrimitives,
       this.kvStore,
       this.eventHook,
       (path) => {
         // TODO: At some point we should remove the data.db exception here
-        return path !== "data.db" && !plugSpacePrimitives.isLikelyHandled(path);
+        return path !== "data.db" &&
+            // Exclude all plug space primitives paths
+            !plugSpaceRemotePrimitives.isLikelyHandled(path) ||
+          // Except federated ones
+          path.startsWith("!");
       },
     );
 
@@ -1108,6 +1113,7 @@ export class Editor {
     const linePrefix = line.text.slice(0, selection.from - line.from);
 
     const results = await this.dispatchAppEvent(eventName, {
+      pageName: this.currentPage!,
       linePrefix,
       pos: selection.from,
     } as CompleteEvent);