+++ web/syscalls/editor.ts
@@ -182,6 +182,10 @@ export function editorSyscalls(editor: Editor): SysCallMapping {
       const cm = vimGetCm(editor.editorView!)!;
       return Vim.handleEx(cm, exCommand);
     },
+    // Sync
+    "editor.syncSpace": () => {
+      return editor.syncService.syncSpace();
+    },
     // Folding
     "editor.fold": () => {
       foldCode(editor.editorView!);